Finish free 30 first minutes parking and corrections
Changes in the function processExitingVehicle in the class ParkingService to not become a recurrent user with a free ticket.
Changes in Data.sql to create database with new table (vehicle) and new attribute (REDUCTION_RATE in ticket table.
Corrections in DBConstants (SAVE_TICKET and GET_TICKET)
Corrections in ticketDAO (functions saveTicket and getTicket)